@@ -18,6 +18,7 @@ import org.springframework.web.client.RestTemplate;
 public class NumericController {
 
 	private final Logger logger = LoggerFactory.getLogger(getClass());
+	//private static final String baseURL = "http://localhost:5005/plusone";
 	private static final String baseURL = "http://node-service:5000/plusone";
 	
 	RestTemplate restTemplate = new RestTemplate();
@@ -50,5 +51,4 @@ public class NumericController {
 			return Integer.parseInt(response);
 		}
 	}
-
 }
